Refactor search controller for consistency and clarity
This reverts commit 7e009fd74a26f9c707fa09d82638a6077a88ffa7.
