+++ api/catalog/api/serializers/media_serializers.py
@@ -3,7 +3,7 @@ from collections import namedtuple
 from rest_framework import serializers
 
 from catalog.api.constants.licenses import LICENSE_GROUPS
-from catalog.api.controllers import search_controller
+from catalog.api.controllers.elasticsearch.stats import get_stats
 from catalog.api.models.media import AbstractMedia
 from catalog.api.serializers.base import BaseModelSerializer
 from catalog.api.serializers.fields import SchemableHyperlinkedIdentityField
@@ -40,6 +40,7 @@ class MediaSearchRequestSerializer(serializers.Serializer):
         "extension",
         "mature",
         "qa",
+        "page",
         "page_size",
     ]
     """
@@ -106,16 +107,22 @@ class MediaSearchRequestSerializer(serializers.Serializer):
         required=False,
         default=False,
     )
+    page = serializers.IntegerField(
+        min_value=1,
+        default=1,
+        help_text="The index of the page of the results to show.",
+    )
     page_size = serializers.IntegerField(
-        label="page_size",
-        help_text="Number of results to return per page.",
-        required=False,
+        min_value=1,
+        max_value=500,
+        default=20,
+        help_text="The number of results to show in one page.",
     )
 
     @staticmethod
     def _truncate(value):
         max_length = 200
-        return value if len(value) <= max_length else value[:max_length]
+        return value[:max_length]
 
     def validate_q(self, value):
         return self._truncate(value)
@@ -366,7 +373,7 @@ def get_search_request_source_serializer(media_type):
 
         _field_attrs = {
             "help_text": make_comma_separated_help_text(
-                search_controller.get_sources(media_type).keys(), "data sources"
+                get_stats(media_type).keys(), "data sources"
             ),
             "required": False,
         }
@@ -384,7 +391,7 @@ def get_search_request_source_serializer(media_type):
         def validate_source_field(value):
             """Checks whether source is a valid source."""
 
-            allowed_sources = list(search_controller.get_sources(media_type).keys())
+            allowed_sources = list(get_stats(media_type).keys())
             sources = value.lower().split(",")
             sources = [source for source in sources if source in allowed_sources]
             value = ",".join(sources)
